Fix Ariane out-of-bound bug

+++ include/Surelog/Expression/ExprBuilder.h
@@ -42,7 +42,7 @@ class ExprBuilder final {
   Value* evalExpr(const FileContent*, NodeId id,
                   ValuedComponentI* instance = nullptr,
                   bool muteErrors = false);
-  Value* fromVpiValue(const std::string& value, unsigned short size);
+  Value* fromVpiValue(const std::string& value, int size);
   Value* fromString(const std::string& value);
   Value* clone(Value* val);
   void seterrorReporting(ErrorContainer* errors, SymbolTable* symbols) {
